GameEngine.c: slight bug fix or workaround?
Only because in 1-step.dump it shows that it forgets about the
origConsts on the next step, which wasn't supposed to happen. I didn't
*expect* it to happen since it also steps forward through `frame`, not
sure what's going on. In any case it doesn't appear to have an effect
on the generated images (which makes sense since while
smallExplosionConsts contains two images, only one is displayed
anyway--which might be a bug?).