cbuild: schedule costliest tasks first
To try to minimize the total real time.

This sorting works on waiting tasks, not all of them, but that should
not be a big problem: if the tasks we start before finishing the
dependency evaluation in the main thread are large ones, that's what
we want, if they are small ones, they will soon be over; if they are
medium ones, bad luck, fair enough, but waiting with scheduling until
the dependency evaluation is finished doesn't look very promising
either.

NOTE: this is inefficient with many waiting tasks! Should use a proper
data structure; sigh: will need CPAN dependency and/or costly
loading.